+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/reducer/ContentStateReducer.kt
@@ -204,6 +204,12 @@ internal object ContentStateReducer {
             is ContentAction.UpdatePermissionHighlightsStateAction -> updateContentState(state, action.sessionId) {
                 it.copy(permissionHighlights = action.highlights)
             }
+            is ContentAction.UpdateAppIntentAction -> updateContentState(state, action.sessionId) {
+                it.copy(appIntent = action.appIntent)
+            }
+            is ContentAction.ConsumeAppIntentAction -> updateContentState(state, action.sessionId) {
+                it.copy(appIntent = null)
+            }
         }
     }
 }