+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/state/AppIntentState.kt
@@ -2,15 +2,17 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.components.browser.session.engine.request
+package mozilla.components.browser.state.state
 
 import android.content.Intent
 
-class LaunchIntentMetadata(
+/**
+ * State keeping track of app intents to launch.
+ *
+ * @param url the URL to launch in an external app.
+ * @param appIntent the [Intent] to launch.
+ */
+data class AppIntentState(
     val url: String,
     val appIntent: Intent?
-) {
-    companion object {
-        val blank = LaunchIntentMetadata("about:blank", null)
-    }
-}
+)